crawlers/requirements.txt
# HTTP Requests and Web Scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
scrapy>=2.11.0

# Vietnamese Text Processing
underthesea>=6.7.0
pyvi>=0.1.1

# Data Processing and Validation
pandas>=2.0.0
numpy>=1.24.0
pydantic>=2.0.0

# File and Environment Handling
python-dotenv>=1.0.0

# Logging and Monitoring
loguru>=0.7.0

# HTTP Client Enhancements
httpx>=0.25.0
aiohttp>=3.8.0

# Text Processing and Utilities
python-dateutil>=2.8.0
pytz>=2023.3

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.11.0

# Development Tools
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0
